refactor(runs): decompose pr.test.ts below 500-line budget (warren-895b)
Extract shared recordingFetch/jsonResponse helpers into pr.test-helpers.ts
and split the PR-merge/URL-parse describe blocks into pr-checks.test.ts.
Remove the grandfathered budget entry; all siblings default-pass < 500.
