+++ scripts/file-size-budgets.json
@@ -9,7 +9,6 @@
 		"src/client/client.ts": 569,
 		"src/db/schema/sqlite.ts": 564,
 		"src/plots/aggregate.ts": 506,
-		"src/runs/pr.test.ts": 604,
 		"src/runs/pr.ts": 659,
 		"src/server/bridges.test.ts": 557,
 		"src/server/handlers/agents.test.ts": 533,
